% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6causal_linear.R
\name{analytic_linear_gaussian_conditining}
\alias{analytic_linear_gaussian_conditining}
\title{Return the mean and the covariance matrix of the conditional distribution of a linear Gaussian SCM}
\usage{
analytic_linear_gaussian_conditining(scm, situation)
}
\arguments{
\item{scm}{A linear Gaussian SCM}

\item{situation}{A list with the following element:
\itemize{
\item condition : either a string that gives an SQL query ( e.g. "select x,y,z from DATA where" )
or a data.table consisting of the valid rows ( e.g. data.table::data.table( x = 0, y = 0))
}}
}
\description{
Return the mean and the covariance matrix of the conditional distribution of a linear Gaussian SCM
}
