% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6causal_examples.R
\docType{data}
\name{backdoor}
\alias{backdoor}
\title{SCM "backdoor" used in the examples.}
\format{
An object of class \code{SCM} (inherits from \code{R6}) of length 40.
}
\usage{
backdoor
}
\description{
Variable z fulfills the back-door criterion for P(y|do(x))
}
\examples{
backdoor
backdoor$plot()
}
\keyword{datasets}
