% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6causal.R
\docType{data}
\name{backdoor_md}
\alias{backdoor_md}
\title{SCM "backdoor_md" used in the examples.}
\format{
An object of class \code{SCM} (inherits from \code{R6}) of length 27.
}
\usage{
backdoor_md
}
\description{
Variable z fulfills the back-door criterion for P(y|do(x)).
Variable z is missing completely at random. The missingness of
variables x and y depend on z.
}
\examples{
backdoor_md
backdoor_md$plot()
}
\keyword{datasets}
