% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6causal_examples.R
\docType{data}
\name{credit}
\alias{credit}
\title{SCM "credit" used in the credit scoring example.}
\format{
An object of class \code{SCM} (inherits from \code{R6}) of length 42.
}
\usage{
credit
}
\description{
Variable \code{default} is the outcome to be predicted
}
\examples{
credit$simulate(100)
summary(credit$simdata)
}
\keyword{datasets}
