% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6causal.R
\docType{data}
\name{frontdoor}
\alias{frontdoor}
\title{SCM "frontdoor" used in the examples.}
\format{
An object of class \code{SCM} (inherits from \code{R6}) of length 21.
}
\usage{
frontdoor
}
\description{
Variable z fulfills the front-door criterion for P(y|do(x))
}
\examples{
frontdoor
frontdoor$plot()
}
\keyword{datasets}
