% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6causal.R
\docType{data}
\name{trapdoor}
\alias{trapdoor}
\title{SCM "trapdoor" used in the examples.}
\format{
An object of class \code{SCM} (inherits from \code{R6}) of length 27.
}
\usage{
trapdoor
}
\description{
Variable z is a trapdoor variable for P(y|do(x))
}
\examples{
trapdoor
trapdoor$plot()
}
\references{
J. Helske, S. Tikka, J. Karvanen (2021). Estimation of causal effects with
small data in the presence of trapdoor variables,
Journal of the Royal Statistical Society Series A, 184(3), 1030-1051,
http://doi.org/10.1111/rssa.12699
}
\keyword{datasets}
